arch: split mcp.rs into protocol, tools, and orchestration (#293)
Split 1,671-line mcp.rs into three focused modules:
- mcp_protocol.rs (260 lines): JSON-RPC types, I/O helpers, bus listener
- mcp_tools.rs (773 lines): all 16 tool implementations + InternalBus
- mcp.rs (577 lines): event loop, request routing, tool registry
