+++ src/app/mod.rs
@@ -15,7 +15,9 @@ pub mod context;
 pub mod graph;
 pub mod jsonrpc;
 pub mod mcp;
+pub mod mcp_protocol;
 pub mod mcp_service;
+pub mod mcp_tools;
 pub mod message;
 pub mod process_builder;
 pub mod schedule;
